+++ addons/l10n_et_base/tests/__init__.py
@@ -7,3 +7,4 @@ from . import test_tax_engine_scope
 from . import test_supported_wht_rates
 from . import test_social_welfare_levy
 from . import test_expense_account_default
+from . import test_tin_on_documents